backend/requirements.txt
# ─── Web Framework ──────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==23.0.0

# ─── Database ───────────────────────────────────────────────────
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# ─── Cache / Message Broker ─────────────────────────────────────
redis==5.2.1
celery==5.4.0
celery[redis]==5.4.0
flower==2.0.1

# ─── Auth & Security ────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pyotp==2.9.0

# ─── HTTP Client (agents / synthetic tests / webhooks) ──────────
httpx==0.28.1
aiohttp==3.11.11

# ─── AI / LLM ───────────────────────────────────────────────────
openai==1.59.3
anthropic==0.40.0
google-generativeai==0.8.3
langchain==0.3.12
langchain-openai==0.2.11
langchain-anthropic==0.3.3

# ─── Time series / ML for baselines ─────────────────────────────
numpy==2.2.1
pandas==2.2.3
scikit-learn==1.6.0
scipy==1.15.0

# ─── Observability / OTel ───────────────────────────────────────
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0

# ─── SSL / Synthetic Tests ──────────────────────────────────────
certifi==2024.12.14
cryptography==44.0.0
playwright==1.49.1

# ─── Email ──────────────────────────────────────────────────────
aiosmtplib==3.0.2
jinja2==3.1.5

# ─── Cloud Integrations ─────────────────────────────────────────
boto3==1.35.93            # AWS
azure-monitor-query==1.4.0 # Azure
google-cloud-monitoring==2.24.0  # GCP

# ─── K8s / VMware ───────────────────────────────────────────────
kubernetes==31.0.0
pyVmomi==8.0.3.0.1

# ─── Utilities ──────────────────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
apscheduler==3.10.4
croniter==5.0.1
pytz==2024.2
arrow==1.3.0
rich==13.9.4
typer==0.15.1
prometheus-client==0.21.1
psutil==6.1.1
paramiko==3.5.0
netmiko==4.5.0
pysnmp==7.1.15
